+++ src/main/java/resources/ChangeCalculationService.java
@@ -18,9 +18,9 @@ public class ChangeCalculationService {
         dollars = new DollarValueConstants();
     }
 
-    public Map<String, Integer> getChange(int changeDue, Map<String, Integer> contents) throws InsufficientFundsException {
-        Map<String, Integer> change = new RegisterContents().getContents();
-        Map<String, Integer> contentsCopy = new HashMap<>();
+    public Map<String, Long> getChange(long changeDue, Map<String, Long> contents) throws InsufficientFundsException {
+        Map<String, Long> change = new RegisterContents().getContents();
+        Map<String, Long> contentsCopy = new HashMap<>();
         contentsCopy.putAll(contents);
 //TODO replace functionality
 //        if (changeDue > getBalance()) {
@@ -33,15 +33,15 @@ public class ChangeCalculationService {
 
         while (changeDue >= dollars.TWENTY && contentsCopy.get(TWENTY) > 0) {
             changeDue -= dollars.TWENTY;
-            change.merge(TWENTY, 1, Integer::sum);
-            contentsCopy.merge(TWENTY, -1, Integer::sum);
+            change.merge(TWENTY, (long) 1, Long::sum);
+            contentsCopy.merge(TWENTY, (long) -1, Long::sum);
         }
 
         while (changeDue >= dollars.TEN && contentsCopy.get(TEN) > 0) {
             if (changeDue >= dollars.TEN + dollars.FIVE || changeDue % 2 == 0 || contentsCopy.get(ONE) > 0) {
                 changeDue -= dollars.TEN;
-                change.merge(TEN, 1, Integer::sum);
-                contentsCopy.merge(TEN, -1, Integer::sum);
+                change.merge(TEN, (long) 1, Long::sum);
+                contentsCopy.merge(TEN, (long) -1, Long::sum);
             } else {
                 break;
             }
@@ -50,8 +50,8 @@ public class ChangeCalculationService {
         while (changeDue >= dollars.FIVE && contentsCopy.get(FIVE) > 0) {
             if (changeDue % 2 == 1 || contentsCopy.get(ONE) > 0 || contentsCopy.get(FIVE) > 1) {
                 changeDue -= dollars.FIVE;
-                change.merge(FIVE, 1, Integer::sum);
-                contentsCopy.merge(FIVE, -1, Integer::sum);
+                change.merge(FIVE, (long) 1, Long::sum);
+                contentsCopy.merge(FIVE, (long) -1, Long::sum);
             } else {
                 break;
             }
@@ -59,14 +59,14 @@ public class ChangeCalculationService {
 
         while (changeDue >= dollars.TWO && contentsCopy.get(TWO) > 0) {
             changeDue -= dollars.TWO;
-            change.merge(TWO, 1, Integer::sum);
-            contentsCopy.merge(TWO, -1, Integer::sum);
+            change.merge(TWO, (long) 1, Long::sum);
+            contentsCopy.merge(TWO, (long) -1, Long::sum);
         }
 
         while (changeDue >= dollars.ONE && contentsCopy.get(ONE) > 0) {
             changeDue -= dollars.ONE;
-            change.merge(ONE, 1, Integer::sum);
-            contentsCopy.merge(ONE, -1, Integer::sum);
+            change.merge(ONE, (long) 1, Long::sum);
+            contentsCopy.merge(ONE, (long) -1, Long::sum);
         }
 
         if (changeDue != 0) {
